requirements2.txt
aiosqlite
fastapi
pydantic
python-dateutil
python-multipart
python-telegram-bot
sqlalchemy[asyncio]
uvicorn[standard]
